Move depends and suggests to the voctomix package, #834296.
Move the depends on gstreamer1.0-tools and the suggests on fbset, python3-pyinotify, gstreamer1.0-libav and rlwrap from the voctomix-core package to the voctomix package, as they are only needed to run some of the scripts and examples from the voctomix package. Thanks to Peter Körner for the  bugreport! (Closes: #834296)