videomixer / swirch a-source to b-source when switching from fullscreen to a composite mode with a new default for a
usually switching from one mode to another keeps a and b sources unchanged. there is however the option to configure defaults for a and b, usually used for preferring slides in the a-source when switching to a composite-mode. this feature left the b-source unchanged which could lead to undesirable behaviour.

in a fullscreen-mix with a=cam2 and b=cam3, cam2 is completely visible, cam3 has no meaning to the user. when the user now switched to a composite-mode with a default, a would become slides and b would 'stay' at cam3, which is unexpected by the user. this feature detects this switch (from fullscreen to a composition) and makes the old a-source the new b-source, before applying the default to the new a-source. with the given example the result would be a=slides and b=cam2 which is much more closer to the expected behaviour to the user